% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsc_pdf.R
\name{qsc_pdf}
\alias{qsc_pdf}
\title{Rmarkdown qsc_pdf template function}
\usage{
qsc_pdf(...)
}
\arguments{
\item{...}{Parameters to pass to pdf_document2().}
}
\description{
See \code{\link[dsreportr]{ds_pdf}} for details.
}
