% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54_pdf.R
\name{u54_pdf}
\alias{u54_pdf}
\title{Rmarkdown u54_pdf template function}
\usage{
u54_pdf(...)
}
\arguments{
\item{...}{Parameters to pass to pdf_document2().}
}
\description{
See \code{\link[dsreportr]{ds_pdf}} for details.
}
